tesseract/requirements.txt
firebase-admin
pytesseract
opencv-python
pillow
pdf2image
python-docx
numpy
